feat(spa): replace sidebar title with ledger switcher
Wire LedgerSwitcher into Sidebar in place of the static 'kea' heading.
Split makeTestApp into a separate test-app.tsx helper so that the global
setup file no longer eagerly loads the route tree, preventing vi.mock()
stale-binding issues in ledger-switcher.test.tsx.
